1.7.031600

+++ exec.loopdisks.php
@@ -37,14 +37,27 @@ function build(){
 	
 	
 	$q=new mysql();
-	$sql="SELECT * FROM loop_disks ORDER BY `size` DESC";$sql="SELECT * FROM loop_disks ORDER BY `size` DESC";
+	$sql="SELECT * FROM loop_disks ORDER BY `size` DESC";
+	
 	$results=$q->QUERY_SQL($sql,"artica_backup");	
-	if(!$q->ok){echo "$q->mysql_error\n";return;}
+	if(!$q->ok){
+		echo "Starting......: Loop disks $q->mysql_error\n";
+		$unix->THREAD_COMMAND_SET($unix->LOCATE_PHP5_BIN()." ".__FILE__);
+		return;
+	}
 	
 	
 	while($ligne=@mysql_fetch_array($results,MYSQL_ASSOC)){
-		$path=$ligne["path"];
+		$path=trim($ligne["path"]);
+		if(is_dir($path)){
+			$oldpath=$path;
+			$path=$path."/".time().".disk";
+			$sql="UPDATE loop_disks SET path='$path' WHERE `path`='$oldpath'";
+			$q->QUERY_SQL($sql,'artica_backup');
+		}
 		$size=$ligne["size"];
+		$maxfds=$ligne["maxfds"];
+		$label=$ligne["disk_name"];
 		writelogs("check $path ($size)",__FUNCTION__,__FILE__,__LINE__);
 		if(!stat_system($path)){
 			writelogs("buil_dd $path ($size)",__FUNCTION__,__FILE__,__LINE__);
@@ -57,26 +70,31 @@ function build(){
 		}
 		if($GetLoops[$path]==null){
 			writelogs("$path no such loop",__FUNCTION__,__FILE__,__LINE__);
-			if(!build_loop($path)){writelogs("$path unable to create loop",__FUNCTION__,__FILE__,__LINE__);continue;}
+			if(!build_loop($path)){writelogs("`$path` unable to create loop",__FUNCTION__,__FILE__,__LINE__);continue;}
+			writelogs("Re-check the loop list...",__FUNCTION__,__FILE__,__LINE__);
 			$GetLoops=GetLoops();
 			if($GetLoops[$path]==null){writelogs("$path no such loop",__FUNCTION__,__FILE__,__LINE__);continue;}
 		}
 		
 		writelogs("$path loop={$GetLoops[$path]}",__FUNCTION__,__FILE__,__LINE__);
 		$sql="UPDATE loop_disks SET loop_dev='{$GetLoops[$path]}' WHERE `path`='$path'";
-		$q=new mysql();
+		
 		$q->QUERY_SQL($sql,'artica_backup');
 		if(!$q->ok){echo "$q->mysql_error\n";continue;}
 		$dev=$GetLoops[$path];
 		echo "Starting......: $path is $dev\n";	
-		if(!ifFileSystem($dev)){if(!mke2fs($dev)){continue;}}
+		if(!ifFileSystem($dev)){if(!mke2fs($dev,$label,$maxfds)){continue;}}
 		$uuid=Getuuid($dev);
 		echo "Starting......: $dev uuid=$uuid\n";
-		if($uuid==null){continue;}		
+		if($uuid==null){continue;}
+
+		
+		
 		$autofs=new autofs();
 		$autofs->uuid=$uuid;
 		$autofs->by_uuid_addmedia($ligne["disk_name"],"auto");
-		shell_exec("/etc/init.d/autofs reload");
+		$nohup=$unix->find_program("nohup");
+		shell_exec("$nohup /etc/init.d/artica-postfix autofs restart >/dev/null 2>&1 &");
 	}	
 	
 	
@@ -87,19 +105,28 @@ function GetLastLo(){
 	return trim(@implode("",$results));
 }
 
-function mke2fs($dev){
+
+
+function mke2fs($dev,$label,$maxfds=0){
 	$debug=$GLOBALS["VERBOSE"];
+	$label_cmd=null;
+	$maxfds_cmd=null;
+	$label=strtolower(str_replace(" ", "_", $label));
+	$label=trim(substr($label, 0,16));
 	$unix=new unix();
 	$mkfs_ext4=$unix->find_program("mkfs.ext4");
+	if($maxfds>0){
+		$maxfds_cmd=" -I 128 -N $maxfds";
+	}
 		
 	if(!is_file($mkfs_ext4)){$mkfs_ext4=$unix->find_program("mkfs.ext3");}
 	
 	if(!$unix->IsExt4()){
 		$mkfs_ext4=$unix->find_program("mkfs.ext3");
 	}	
-	
+	if($label<>null){$label_cmd=" -L $label";}
 	echo "Starting......: $dev formatting...\n";		
-	$cmd="$mkfs_ext4 -q $dev 2>&1";
+	$cmd="$mkfs_ext4 $label_cmd$maxfds_cmd -q $dev 2>&1";
 	exec($cmd,$results);
 	if($debug){echo "mke2fs($dev) -> $cmd ". count($results)." rows\n";}	
 	if($debug){while (list ($num, $line) = each ($results)){echo "mke2fs() -> $line\n";}}
@@ -125,6 +152,7 @@ function build_dd($path,$size){
 	$dd=$unix->find_program("dd");
 	$size=$size*1024;
 	$cmd="$dd if=/dev/zero of=$path bs=1024 count=$size 2>&1";
+	if($GLOBALS["VERBOSE"]){echo "$cmd\n";}
 	exec($cmd,$results);
 	echo "build_dd() $cmd ". count($results)." rows\n";
 	while (list ($num, $ligne) = each ($results) ){echo "build_dd() $ligne\n";}
@@ -135,12 +163,13 @@ function build_dd($path,$size){
 function build_loop($path){
 	$loop_free=GetLastLo();
 	$cmd="{$GLOBALS["losetup"]} $loop_free $path 2>&1";
+	if($GLOBALS["VERBOSE"]){echo "$cmd\n";}
 	exec($cmd,$results);
-	echo "build_loop() $cmd ". count($results)." rows\n";
+	if($GLOBALS["VERBOSE"]){echo "build_loop() $cmd ". count($results)." rows\n";}
 	while (list ($num, $ligne) = each ($results) ){echo "build_loop() $ligne\n";}
 	$GetLoops=GetLoops();
 	if($GetLoops[$path]<>null){
-		echo "build_loop() done {$GetLoops[$path]}\n";
+		if($GLOBALS["VERBOSE"]){echo "build_loop() done {$GetLoops[$path]}\n";}
 		return true;	
 	}	
 	return false;
@@ -149,7 +178,11 @@ function build_loop($path){
 function GetLoops(){
 	$cmd="{$GLOBALS["losetup"]} -a 2>&1";
 	exec($cmd,$results);	
-	while (list ($num, $ligne) = each ($results) ){if(preg_match("#^(.+?):.+?\((.+?)\)#",$ligne,$re)){$array[trim($re[2])]=trim($re[1]);}}	
+	while (list ($num, $ligne) = each ($results) ){
+		if(preg_match("#^(.+?):.+?\((.+?)\)#",$ligne,$re)){
+			$array[trim($re[2])]=trim($re[1]);
+		}
+	}	
 	return $array;
 	
 }
@@ -161,8 +194,10 @@ function remove($path){
 	$q=new mysql();
 	$ligne=mysql_fetch_array($q->QUERY_SQL($sql,"artica_backup"));
 	$disk_name=$ligne["disk_name"];
+	$loop_dev=$ligne["loop_dev"];
 	$GetLoops=GetLoops();
 	$dev=$GetLoops[$path];
+	if($dev==null){$dev=$loop_dev;}
 	$uuid=Getuuid($dev);
 	if($dev<>null){
 		echo "Starting......: $dev umounting...\n";
@@ -172,6 +207,8 @@ function remove($path){
 		while (list ($num, $ligne) = each ($results) ){echo "Starting......: $dev $ligne\n";}
 		
 	}
+	
+	
 	$results=array();
 	if($uuid<>null){
 		echo "Starting......: $dev disconnect $uuid...$disk_name\n";
@@ -179,17 +216,23 @@ function remove($path){
 		$autofs->uuid=$uuid;
 		$autofs->by_uuid_removemedia($disk_name,"auto");		
 	}
-	echo "Starting......: $dev remove media\n";
-	$cmd="{$GLOBALS["losetup"]} -d $dev 2>&1";
-	exec($cmd,$results);	
-	while (list ($num, $ligne) = each ($results) ){echo "Starting......: $dev $ligne\n";}	
-	echo "Starting......: $dev remove file\n";
-	shell_exec("/bin/rm -f $path");
+	
+	if($dev<>null){
+		echo "Starting......: dev:`$dev` remove media\n";
+		$cmd="{$GLOBALS["losetup"]} -d $dev 2>&1";
+		exec($cmd,$results);	
+		while (list ($num, $ligne) = each ($results) ){echo "Starting......: $dev $ligne\n";}	
+		if(is_file($path)){
+			echo "Starting......: $dev remove file\n";
+			shell_exec("/bin/rm -f $path");
+		}
+	}
 	echo "Starting......: $dev remove entry in database\n";
 	$sql="DELETE FROM loop_disks WHERE `path`='$path'";
 	$q->QUERY_SQL($sql,"artica_backup");
 	echo "Starting......: $dev removed\n";
-	shell_exec("/etc/init.d/autofs restart");
+	$nohup=$unix->find_program("nohup");
+	shell_exec("$nohup /etc/init.d/autofs restart >/dev/null 2>&1 &");
 	
 }
 
@@ -219,7 +262,10 @@ function ifFileSystem($dev){
 
 
 function stat_system($path){
-	exec("stat -f $path -c %b 2>&1",$results);
+	$unix=new unix();
+	$stat=$unix->find_program("stat");
+	if($GLOBALS["VERBOSE"]){echo "stat -f $path -c %b 2>&1\n";}
+	exec("$stat -f $path -c %b 2>&1",$results);
 	$line=trim(@implode("",$results));
 	if(preg_match("#^[0-9]+#",$line,$results)){return true;}
 	return false;
